Move exported internal types to their own file
This makes it clear which internal types are publicly exported.
